Working example to show stereo Texture.
Not sure exactly what is happening, but this is a good place to snapshot.

If I create a Texture with RT usage, and then GetSurfaceLevel on that, then the StretchRect still shows a stretched image, but when I look at the Texture with Pix, I see stereo.

I'm not sure what this means, but want to be able to get back here if I bust it.