Variant that seems to work.
This adds a StretchRect back to the main screen, of the saved stereo backbuffer, gotten from the Texture, not the Surface.

But then copying the Surface back to backbuffer, actually shows both sides in stereo.

So, it looks like this works to create a Surface that is stereo.

I'm not positive this is correct, but when I cross-eye view the end result, it is stereo, with right eye on left.