Add HR macros to simplify error checking.
Using Frank Luna's HR macro to simplify error checking and handling, since we never expect errors, and can just throw up the dialog with info.

This is set to run in both Release and Debug, and changes the linking slightly to handle the use of dxerr.lib.

This dramatically improves the code by pulling out the MessageBox calls, but still handles errors.

Since this changes so many lines of code, this also includes a reformat of the code, which changes even more lines by changing whitespace, but making everything consistent.